PM2020-22: rework facets
