+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/query_builders.yml
@@ -798,7 +798,7 @@ services:
             - '@pim_enrich.query.product_and_product_model_document_type_facet_query'
 
     pim_enrich.query.product_and_product_model_document_type_facet_query:
-        class: 'Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Aggregation\ProductAndProductsModelDocumentTypeFacetQuery'
+        class: 'Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Facet\ProductAndProductsModelDocumentTypeFacetQuery'
 
     pim_enrich.query.product_and_product_model_document_type_facet_factory:
-        class: 'Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Aggregation\ProductAndProductsModelDocumentTypeFacetFactory'
+        class: 'Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Facet\ProductAndProductsModelDocumentTypeFacetFactory'
